backend/pyproject.toml
[tool.black]
line-length = 100
target-version = ["py37"]

[tool.isort]
dont-skip = ["__init__.py"]
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = "backend, tests"
known_third_party = "itsdangerous, pyotp, quart, toml, zxcvbn"
line_length = 100
multi_line_output = 3
no_lines_before = "LOCALFOLDER"
order_by_type = false

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["pgjones <philip.graham.jones@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
hypercorn = { git = "https://gitlab.com/pgjones/hypercorn.git", branch = "master" }
quart-trio = "*"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]
black = { version = "*", allows-prereleases = true }
pytest = "*"
isort = "*"
flake8 = "*"
pep8-naming = "*"
flake8-print = "*"
mypy = "*"
pytest-trio = "*"
bandit = "*"

[tool.poetry.scripts]
format = "scripts:format"
reformat = "scripts:reformat"
lint = "scripts:lint"
start = "scripts:start"
test = "scripts:test"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
